+++ providers/sync/src/utils/assets.utils.ts
@@ -15,7 +15,7 @@ export const getSlidesLocalImages = async ({deck}: {deck: Deck | undefined}): Pr
 };
 
 export const getSlidesOnlineImages = async ({deck}: {deck: Deck | undefined}): Promise<UserAsset[]> => {
-  return getAssets<UserAsset>({elementIds: deck?.data?.slides, assets: getSlideOnlineImages, selector: slideSelector});
+  return getAssets<UserAsset>({elementIds: deck?.data?.slides, assets: getOnlineImages, selector: slideSelector});
 };
 
 export const getSlidesLocalCharts = async ({deck}: {deck: Deck | undefined}): Promise<UserAsset[]> => {
@@ -35,7 +35,7 @@ export const getParagraphsLocalImages = async ({doc}: {doc: Doc | undefined}): P
 };
 
 export const getParagraphsOnlineImages = async ({doc}: {doc: Doc | undefined}): Promise<UserAsset[]> => {
-  return getAssets<UserAsset>({elementIds: doc?.data?.paragraphs, assets: getSlideOnlineImages, selector: paragraphSelector});
+  return getAssets<UserAsset>({elementIds: doc?.data?.paragraphs, assets: getOnlineImages, selector: paragraphSelector});
 };
 
 const slideSelector = (id: string) => {
@@ -105,25 +105,35 @@ const getLocalImages = async ({selector}: {selector: string}): Promise<UserAsset
   });
 
   // Read files from idb - preserve keys (the file name might not match the key since we `encodeFilename` with offline providers when we upload the images)
-  const files: {key: string; blob: File}[] = await Promise.all(
-    list.map(async ({imgSrc}: HTMLDeckgoLazyImgElement) => {
-      const blob: File = await get<File>(imgSrc);
+  const toFile = async ({img, attr}: {img: HTMLDeckgoLazyImgElement; attr: 'img-src' | 'data-src'}) => {
+    const key: string = img.getAttribute(attr);
 
-      return {
-        key: imgSrc,
-        blob
-      };
-    })
+    const blob: File = await get<File>(key);
+
+    return {
+      key,
+      blob
+    };
+  };
+
+  const imgFiles: {key: string; blob: File}[] = await Promise.all(
+    list.map((img: HTMLDeckgoLazyImgElement) => toFile({img, attr: 'img-src'}))
   );
 
-  return files.map(({key, blob}: {key: string; blob: File}) => ({
+  const dataFiles: {key: string; blob: File}[] = await Promise.all(
+    list
+      .filter((element: HTMLDeckgoLazyImgElement) => element.hasAttribute('data-src'))
+      .map((img: HTMLDeckgoLazyImgElement) => toFile({img, attr: 'data-src'}))
+  );
+
+  return [...imgFiles, ...dataFiles].map(({key, blob}: {key: string; blob: File}) => ({
     key,
     blob,
     type: 'local'
   }));
 };
 
-const getSlideOnlineImages = async ({selector}: {selector: string}): Promise<UserAsset[] | undefined> => {
+const getOnlineImages = async ({selector}: {selector: string}): Promise<UserAsset[] | undefined> => {
   const imgs: HTMLDeckgoLazyImgElement[] | undefined = getImages({selector});
 
   if (!imgs || imgs.length <= 0) {
@@ -135,11 +145,19 @@ const getSlideOnlineImages = async ({selector}: {selector: string}): Promise<Use
     return !isLocalImage(img);
   });
 
-  const promises: Promise<UserAsset | undefined>[] = list.map(({imgSrc}: HTMLDeckgoLazyImgElement) =>
-    getUserAsset({url: imgSrc, type: 'images'})
+  const toFile = ({img, attr}: {img: HTMLDeckgoLazyImgElement; attr: 'img-src' | 'data-src'}) =>
+    getUserAsset({url: img.getAttribute(attr), type: 'images'});
+
+  const imgFiles: (UserAsset | undefined)[] = await Promise.all(
+    list.map((img: HTMLDeckgoLazyImgElement) => toFile({img, attr: 'img-src'}))
+  );
+  const dataFiles: (UserAsset | undefined)[] = await Promise.all(
+    list
+      .filter((element: HTMLDeckgoLazyImgElement) => element.hasAttribute('data-src'))
+      .map((img: HTMLDeckgoLazyImgElement) => toFile({img, attr: 'data-src'}))
   );
 
-  return (await Promise.all(promises)).filter((asset: UserAsset | undefined) => asset !== undefined);
+  return [...imgFiles, ...dataFiles].filter((asset: UserAsset | undefined) => asset !== undefined);
 };
 
 const getImages = ({selector}: {selector: string}): HTMLDeckgoLazyImgElement[] | undefined => {
@@ -147,6 +165,10 @@ const getImages = ({selector}: {selector: string}): HTMLDeckgoLazyImgElement[] |
 
   if (!element) {
     return undefined;
+  }
+
+  if (element.nodeName.toLowerCase() === 'deckgo-lazy-img') {
+    return [element as HTMLDeckgoLazyImgElement];
   }
 
   const imgs: NodeListOf<HTMLDeckgoLazyImgElement> = element.querySelectorAll('deckgo-lazy-img');